requirements.txt
torch>=1.8.0
torchvision>=0.9.0
Pillow>=8.0.0
matplotlib>=3.3.0
numpy>=1.18.5
